refactor: restructure method validation package and update examples
